move websocket connection to specific hook
