+++ src/app/App.tsx
@@ -5,24 +5,17 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import React, { FunctionComponent, useCallback, useEffect, useState } from 'react';
+import React, { FunctionComponent, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useMatch, useNavigate } from 'react-router';
-import {
-    AuthenticationRouter,
-    CardErrorBoundary,
-    initializeAuthenticationProd,
-    useSnackMessage,
-} from '@gridsuite/commons-ui';
-import { connectNotificationsWsUpdateConfig } from '../utils/rest-api';
+import { AuthenticationRouter, CardErrorBoundary, initializeAuthenticationProd } from '@gridsuite/commons-ui';
 import AppTopBar, { AppTopBarProps } from '../features/top-bar/components/app-top-bar';
-import ReconnectingWebSocket from 'reconnecting-websocket';
 import { getErrorMessage } from '../shared/lib/error';
 import { AppRouter } from 'app/router/AppRouter';
-import { AppState } from './store/reducer';
 import { selectAuthentication } from 'features/authentication/store/authentication.selectors';
-import { invalidateConfigQueries } from 'shared/api/config-api/config-api';
 import { fetchIdpSettings } from 'shared/config/idp-settings';
+import { AppState } from './store/store.type';
+import { useAppParametersInvalidationListener } from 'features/app-parameters/hooks/use-app-parameters-invalidation-listener';
 
 const App: FunctionComponent = () => {
     const user = useSelector((state: AppState) => selectAuthentication(state).user);
@@ -42,20 +35,6 @@ const App: FunctionComponent = () => {
     const dispatch = useDispatch();
 
     const location = useLocation();
-
-    const connectNotificationsUpdateConfig = useCallback((): ReconnectingWebSocket => {
-        const ws = connectNotificationsWsUpdateConfig();
-        ws.onmessage = function (event) {
-            let eventData = JSON.parse(event.data);
-            if (eventData.headers?.parameterName) {
-                invalidateConfigQueries(dispatch, eventData.headers?.parameterName);
-            }
-        };
-        ws.onerror = function (event) {
-            console.error('Unexpected Notification WebSocket error', event);
-        };
-        return ws;
-    }, []);
 
     // Can't use lazy initializer because useMatch is a hook
     const [initialMatchSilentRenewCallbackUrl] = useState(
@@ -93,12 +72,7 @@ const App: FunctionComponent = () => {
         // Note: dispatch and initialMatchSilentRenewCallbackUrl won't change
     }, [initialMatchSigninCallbackUrl, initialMatchSilentRenewCallbackUrl, dispatch]);
 
-    useEffect(() => {
-        if (user !== null) {
-            const ws = connectNotificationsUpdateConfig();
-            return () => ws.close();
-        }
-    }, [user, connectNotificationsUpdateConfig]);
+    useAppParametersInvalidationListener();
 
     return (
         <>
