+++ src/app/store/reducer.ts
@@ -6,17 +6,8 @@
  */
 
 import { combineReducers } from '@reduxjs/toolkit';
-import { AuthenticationActions } from '@gridsuite/commons-ui';
-import { AuthenticationState } from 'features/authentication/store/authentication.type';
 import { authenticationReducer } from 'features/authentication/store/authentication.slice';
 import { baseApi } from 'shared/api/base-api';
-
-export type AppState = {
-    authentication: AuthenticationState;
-    [baseApi.reducerPath]: ReturnType<typeof baseApi.reducer>;
-};
-
-export type Actions = AuthenticationActions;
 
 export const reducer = combineReducers({
     authentication: authenticationReducer,
