+++ src/shared/api/base-api.ts
@@ -1,14 +1,13 @@
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react';
-import { AppState } from 'app/store/reducer';
 
 export const baseApi = createApi({
     reducerPath: 'api',
     baseQuery: fetchBaseQuery({
         baseUrl: import.meta.env.VITE_API_GATEWAY,
         prepareHeaders: (headers, { getState }) => {
-            const state = getState() as AppState;
+            const state = getState() as any;
 
-            const token = state.authentication?.user?.id_token; // adapte selon ton store
+            const token = state?.authentication?.user?.id_token;
 
             if (token) {
                 headers.set('Authorization', `Bearer ${token}`);
