+++ src/shared/api/ws/config-ws.ts
@@ -5,11 +5,10 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { AppState } from 'app/store/reducer';
-import { APP_NAME } from '../app/config/config';
 import ReconnectingWebSocket from 'reconnecting-websocket';
 import { store } from 'app/store/store';
 import { selectAuthentication } from 'features/authentication/store/authentication.selectors';
+import { APP_NAME } from 'app/config/config';
 
 export interface ErrorWithStatus extends Error {
     status?: number;
@@ -19,20 +18,10 @@ export type Url = string | URL;
 export type InitRequest = Partial<RequestInit>;
 export type Token = string;
 
-// const PREFIX_USER_ADMIN_SERVER_QUERIES = `${import.meta.env.VITE_API_GATEWAY}/user-admin`;
-
-// If you want to use user-admin-server in dev mode you must avoid passing through gateway
-// and use the user-admin-server directly. SetupProxy should allow this.
-// const PREFIX_USER_ADMIN_SERVER_QUERIES =
-//     process.env.REACT_APP_API_PREFIX +
-//     (process.env.REACT_APP_USE_AUTHENTICATION === 'true'
-//         ? `${import.meta.env.VITE_API_GATEWAY}/user-admin`
-//         : process.env.REACT_APP_USER_ADMIN_URI);
-
 const PREFIX_CONFIG_NOTIFICATION_WS = `${import.meta.env.VITE_WS_GATEWAY}/config-notification`;
 
 function getToken(): Token | null {
-    const state: AppState = store.getState();
+    const state = store.getState();
     return selectAuthentication(state).user?.id_token ?? null;
 }
